requirements.txt
# Core (PyTorch: install a torch build matching your CUDA from https://pytorch.org )
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
torch-ema

# Config & training
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
omegaconf
rootutils
rich
pre-commit
pytest

# Geospatial / weather data stack
numpy
pandas
xarray
dask
netcdf4
h5netcdf
gcsfs
arrow
einops
scipy
timm

# Plotting (metrics / viz utilities)
matplotlib
seaborn
cartopy

# Optional: only if you use GCS in result saving / prepare scripts
# google-cloud-storage
